requirements.txt
playwright
beautifulsoup4
requests
fpdf
tqdm
python-dotenv
psycopg2-binary
PyPDF2
pytesseract
Pillow
pdf2image
